moved resource check func to node.py
